made client configurable for sync/asyc
